+++ src/frequenz/channels/util/_merge_named.py
@@ -7,7 +7,8 @@ import asyncio
 from collections import deque
 from typing import Any, Deque, Set, Tuple
 
-from .._base_classes import ChannelClosedError, Receiver, T
+from .._base_classes import Receiver, T
+from .._exceptions import ReceiverStoppedError
 
 
 class MergeNamed(Receiver[Tuple[str, T]]):
@@ -42,21 +43,28 @@ class MergeNamed(Receiver[Tuple[str, T]]):
         await asyncio.gather(*self._pending, return_exceptions=True)
         self._pending = set()
 
-    async def ready(self) -> None:
-        """Wait until there's a message in any of the channels.
+    async def ready(self) -> bool:
+        """Wait until the receiver is ready with a value or an error.
 
-        Raises:
-            ChannelClosedError: when all the channels are closed.
+        Once a call to `ready()` has finished, the value should be read with
+        a call to `consume()` (`receive()` or iterated over). The receiver will
+        remain ready (this method will return immediately) until it is
+        consumed.
+
+        Returns:
+            Whether the receiver is still active.
         """
         # we use a while loop to continue to wait for new data, in case the
         # previous `wait` completed because a channel was closed.
         while True:
             # if there are messages waiting to be consumed, return immediately.
             if len(self._results) > 0:
-                return
+                return True
 
+            # if there are no more pending receivers, we return immediately.
             if len(self._pending) == 0:
-                raise ChannelClosedError()
+                return False
+
             done, self._pending = await asyncio.wait(
                 self._pending, return_when=asyncio.FIRST_COMPLETED
             )
@@ -75,12 +83,16 @@ class MergeNamed(Receiver[Tuple[str, T]]):
     def consume(self) -> Tuple[str, T]:
         """Return the latest value once `ready` is complete.
 
-        Raises:
-            EOFError: When called before a call to `ready()` finishes.
-
         Returns:
-            The next value that was received, along with its name.
+            The next key, value that was received.
+
+        Raises:
+            ReceiverStoppedError: if the receiver stopped producing messages.
+            ReceiverError: if there is some problem with the receiver.
         """
+        if not self._results and not self._pending:
+            raise ReceiverStoppedError(self)
+
         assert self._results, "calls to `consume()` must be follow a call to `ready()`"
 
         return self._results.popleft()